Matching progress in WaterGun.
Mainly matching work in TNozzleBase::emit, TNozzleDeform::emit and TNozzleTrigger::emit.

Improved matches in various other functions in the TU.

+++ src/Player/ModelWaterManager.cpp
@@ -277,7 +277,7 @@ void TModelWaterManager::makeEmit(const TWaterEmitInfo& param_1)
 	unk2514[mParticleCount] = nullptr;
 }
 
-int TModelWaterManager::emitRequest(const TWaterEmitInfo& param_1)
+u8 TModelWaterManager::emitRequest(const TWaterEmitInfo& param_1)
 {
 	int particlesToSpawn = param_1.mNum.get();
 	if (particlesToSpawn == 0)
